+++ pages/faq.php
@@ -12,7 +12,7 @@
 		</li>
 		<li><b>Wie kann ich das Warenkorb Symbol in meinem Template einbinden?</b><br>
 			<?= rex_string::highlight(<<<'EOF'
-    print '<a href="'. rex_getUrl(rex_config::get('d2u_courses', 'article_id_shopping_cart')) .'" class="cart_link">';
+    print '<a href="'. rex_getUrl((int) rex_config::get('d2u_courses', 'article_id_shopping_cart')) .'" class="cart_link">';
     print '<div id="cart_symbol" class="desktop-inner">';
     print '<img src="'. rex_url::addonAssets('d2u_courses', 'cart_only.png') .'" alt="'.
     	rex_article::get(rex_config::get('d2u_courses', 'article_id_shopping_cart', 0))->getName() .'">';
